UserInputSystem wasd->moveables, recording clicks and double clicks, PhysicsSystem started.

+++ lib/SXNGN/cpp/Database.cpp
@@ -1,8 +1,8 @@
-#include "..\headers\Database.h"
+#include <Database.h>
 #include <stdio.h>
 
-int SXNGN::Database::scale_;
-int SXNGN::Database::max_fps_;
+Uint8 SXNGN::Database::scale_;
+Uint8 SXNGN::Database::max_fps_;
 std::shared_ptr<Coordinator> SXNGN::Database::ecs_coordinator_;
 
 SXNGN::Database::Database()
@@ -13,12 +13,12 @@ SXNGN::Database::Database()
 
 }
 
-int SXNGN::Database::get_scale()
+Uint8 SXNGN::Database::get_scale()
 {
 	return scale_;
 }
 
-void SXNGN::Database::set_scale(int new_scale)
+void SXNGN::Database::set_scale(Uint8 new_scale)
 {
 	scale_ = new_scale;
 }
@@ -42,17 +42,17 @@ void SXNGN::Database::increase_scale()
 }
 
 
-int SXNGN::Database::get_max_fps()
+Uint8 SXNGN::Database::get_max_fps()
 {
 	return max_fps_;
  }
 
-void SXNGN::Database::set_max_fps(int new_fps)
+void SXNGN::Database::set_max_fps(Uint8 new_fps)
 {
 	max_fps_ = new_fps;
  }
 
-int  SXNGN::Database::get_screen_ticks_per_frame()
+Uint8  SXNGN::Database::get_screen_ticks_per_frame()
 {
 	if (max_fps_ == 0)
 	{
